Conflicts:
	.gitignore
	configs/common/core_sim.hal
	src/Makefile
	src/emc/ini/emcIniFile.hh
	src/emc/ini/iniaxis.cc
	src/emc/ini/inihal.cc
	src/emc/ini/initraj.cc
	src/emc/motion/command.c
	src/emc/motion/control.c
	src/emc/motion/motion.c
	src/emc/motion/motion.h
	src/emc/nml_intf/emc.hh
	src/emc/nml_intf/emcglb.c
	src/emc/nml_intf/emcglb.h
	src/emc/nml_intf/interpl.cc
	src/emc/task/emccanon.cc
	src/emc/task/emctaskmain.cc
	src/emc/task/taskintf.cc
	src/emc/tp/tc_types.h
	src/emc/usr_intf/axis/scripts/axis.py
	src/emc/usr_intf/axis/scripts/linuxcnctop.py
	src/hal/components/Submakefile

+++ src/emc/nml_intf/interpl.hh
@@ -17,6 +17,8 @@
 #ifndef INTERP_LIST_HH
 #define INTERP_LIST_HH
 
+#include <stdint.h>
+
 #define MAX_NML_COMMAND_SIZE 1000
 
 // these go on the interp list
@@ -26,21 +28,21 @@ struct NML_INTERP_LIST_NODE {
     int remap_level;            // remap_level for interp._setup.remap_level
 
     union _dummy_union {
-	int i;
-	long l;
+	int32_t i;
+	int32_t l;
 	double d;
 	float f;
-	long long ll;
+	int64_t ll;
 	long double ld;
     } dummy;			// paranoid alignment variable.
 
     union _command_union {
 	char commandbuf[MAX_NML_COMMAND_SIZE];	// the NML command;
-	int i;
-	long l;
+	int32_t i;
+	int32_t l;
 	double d;
 	float f;
-	long long ll;
+	int64_t ll;
 	long double ld;
     } command;
 };
